+++ packages/abjad-convert/src/index.ts
@@ -2,6 +2,7 @@ import { getConverter } from './converterFactory'
 import { Abjad } from './types'
 import { arabic } from './scripts/arabic/letters'
 import { imperialAramaic } from './scripts/imperialAramaic/letters'
+import { ipa } from './scripts/ipa/letters'
 import { oldSouthArabian } from './scripts/oldSouthArabian/letters'
 import { phoenician } from './scripts/phoenician/letters'
 import { syriac } from './scripts/syriac/letters'
@@ -24,6 +25,8 @@ export function getLetters(script: Abjad): readonly string[] {
 		return arabic.letters
 	case Abjad.ImperialAramaic:
 		return imperialAramaic.letters
+	case Abjad.IPA:
+		return ipa.letters
 	case Abjad.OldSouthArabian:
 		return oldSouthArabian.letters
 	case Abjad.Phoenician: