+++ packages/abjad-web/src/components/OnScreenKeyboard.tsx
@@ -10,15 +10,26 @@ type Props = {
 const OnScreenKeyboard: FC<Props> = ({ letters, onClickLetter, onBackSpace, onClearAll }) => {
 	return <div>
 		{letters.map((letter, index) =>
-			<button key={`btn_letter${index}`} onClick={() => onClickLetter(letter)}>{letter}</button>)
+			<button
+				key={`btn_letter${index}`}
+				style={{ marginLeft: '1px', fontSize: '18px' }}
+				onClick={() => onClickLetter(letter)}>
+				{letter}
+			</button>)
 		}
-		<button key={`btn_space`} style={{ marginLeft: '10px', backgroundColor: 'lightsteelblue' }}
+		<button
+			key={`btn_space`}
+			style={{ marginLeft: '10px', backgroundColor: 'lightsteelblue' }}
 			onClick={() => onClickLetter(' ')}>␣
 		</button>
-		<button key={`btn_backSpace`} style={{ marginLeft: '2px', backgroundColor: 'lightsteelblue' }}
+		<button
+			key={`btn_backSpace`}
+			style={{ marginLeft: '2px', backgroundColor: 'lightsteelblue' }}
 			onClick={onBackSpace}>&#x232B;
 		</button>
-		<button key={`btn_clearAll`} style={{ marginLeft: '2px', backgroundColor: 'lightsteelblue' }}
+		<button
+			key={`btn_clearAll`}
+			style={{ marginLeft: '2px', backgroundColor: 'lightsteelblue' }}
 			onClick={onClearAll}>❌
 		</button>
 	</div>